Microservices/auth/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pydantic>=2
pydantic-settings>=2
passlib[bcrypt]
python-jose[cryptography]
emails
jinja2
PyJWT
email-validator
psycopg2-binary
python-multipart
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
pytest